renamed ErrorBoundary directory
